Relocate Scripts and Themes API to Api/ namespace
