+++ src/BigCommerce/Api/PriceLists/PriceListsApi.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace BigCommerce\ApiV3\PriceLists;
+namespace BigCommerce\ApiV3\Api\PriceLists;
 
 use BigCommerce\ApiV3\Api\Generic\ResourceApi;
 use BigCommerce\ApiV3\ResponseModels\PriceList\PriceListResponse;
